Pass information using a strongly typed model
Create the model, `AboutModel`. Change the `AboutController` to create
an instance of the `AboutModel` class and to pass that instance in the
`View` constructor.

Add a directive to `AboutView.cshtml` instructing the Razor engine that
the type of the `Model` property of this view is
`OdeToFood.Models.AboutModel`. Change the HTML to display the two
properties of this model:

* `Name` (Mr. Wizard)
* `Location` (Houston, TX)