Changed name parameter to optional
This commit changes the default value of name parameter for routes like
"/Cuisine/french" to be optional (`UrlParameter.Optional`). This choice,
slightly different from the previous commit supplying an empty string,
informs the MVC framework that this parameter is **optional**. (It does
not actually supply a value if it is missing (although it actually may -
I just do not know enough)).